Restructured the whole app, changed file naming convention from PascalCase to kebab-case.
