+++ src/components/footer/footer.component.css
@@ -1,0 +1,11 @@
+.footer{
+	display: flex;
+	flex-direction: column;
+	align-items: center;
+	align-content: center;
+	border-top: 1px solid;
+}
+.footer h5, .footer h6{
+	margin: 10px 0;
+}
+
