+++ src/components/navbar/navbar.component.js
@@ -1,6 +1,6 @@
 import React, { useContext } from 'react';
-import { ThemeContext } from './ThemeProvider'
-
+import { ThemeContext } from '../theme-provider/theme-provider.component';
+import './navbar.component.css';
 
 export default function Navbar(){
 	const state = useContext(ThemeContext)
