+++ src/components/section-about/section-about.component.css
@@ -1,0 +1,133 @@
+#about-me{
+	padding: 0;
+	margin-bottom: 50px !important;
+}
+
+@media only screen and (min-width: 300px) {
+    .section-2{
+		position: relative;
+		min-height: 1600px;
+		height: auto;
+		transition: all 0.3s ease;
+		display: flex;
+		flex-direction: column;
+		align-items: center;
+	}
+	.section-slant {
+		border-top: 100px solid transparent;
+		border-right: 99vw solid rgb(15, 15, 15);
+		width: 100%;
+		height: 100px;
+		position: absolute;
+		top: -100px;
+	}
+	.section-2-content-container{
+		margin-top: 100px;
+	}
+	
+	.my-photo, #about-me{
+		margin-right: 16px;
+	}
+	.my-photo {
+		width: 225px;
+		height: 225px;
+		border-radius: 50%;
+		box-shadow: 2px 2px 10px antiquewhite;
+	}
+	.section-2 h2{
+		margin: 25px auto;
+		font-weight: 300;	
+		font-size: 18px;
+		text-align: center;
+		width: 250px;
+		height: 300px;
+		position: relative;
+		padding: 20px;
+		display: flex;
+		flex-direction: column;
+		justify-content: center;
+	}
+
+	.contact-me-container{
+		margin: 20px auto;
+	}
+	.contact-me-btn{
+		font-weight: 400;
+		border: 1px solid #00d8ff;
+		border-radius: 5px;
+		padding: 10px 20px;
+		font-size: 20px;		
+	}
+	.contact-me-btn:hover{
+		color: black !important;
+		background-color: #00d8ff;
+	}
+}
+@media only screen and (min-width: 425px) {
+    .my-photo{
+		width: 200px;
+		height: 200px;
+	}
+}
+
+
+@media only screen and (min-width: 768px) {
+	.section-2 {
+		padding-bottom: 100px;
+	}
+	.section-2 h2 {
+		font-size: 20px;
+		width: 300px;
+		height: 350px;
+	}
+	.contact-me-container {
+		margin: 30px auto;
+	}
+}
+
+@media only screen and (min-width: 1024px) {
+	.my-photo{
+		width: 300px;
+		height: 300px;
+		margin-bottom: 50px;
+	}
+}
+
+@media only screen and (min-width: 1440px){
+	.section-2-content-container{
+		flex-direction: row;
+	}
+	.about-me-text-container{
+		display: flex;
+		justify-content: center;
+	}
+	.section-2{
+		padding-bottom: 100px;
+		min-height: 1200px;
+	}
+	.section-2 h2 {
+		font-size: 24px;
+		width: 300px;
+		height: 450px;
+		margin: 0 15px;
+	}
+}
+
+@media only screen and (min-width: 1440px){
+    .section-2{
+        min-height: 800px;
+	}
+	.section-2-content-container{
+        display: flex;
+		flex-direction: row;
+	}
+}
+
+@media only screen and (min-width: 1800px){
+    .section-2 h2 {
+        width: 350px;
+        height: 450px;
+        margin: 0 50px;
+        padding: 0 35px;
+    }
+}
