+++ src/components/section-contact/section-contact.component.css
@@ -1,0 +1,31 @@
+input, textarea{
+	color: rgb(250, 235, 215);
+}
+
+@media only screen and (min-width: 300px){
+    .section-5{
+        position: relative;
+        padding: 50px 0 100px;
+        display: flex;
+        flex-direction: column;
+        align-items: center;
+        justify-content: center;
+        margin: 0 0 -80px 0 !important;
+        transition: all 0.3s ease;
+    }
+    
+    .section-5 .form-box{
+        position: relative;
+        display: flex;
+        flex-direction: column;
+        margin-bottom: 20px;
+        margin-top: 40px;
+    }
+}
+
+@media only screen and (min-width: 4000px){
+	.section-5{
+		margin-bottom: -150px !important;
+	}
+	
+}
