+++ src/components/section-home/section-home.component.css
@@ -1,0 +1,148 @@
+@media only screen and (min-width: 300px){
+	.section-1{
+		margin-top: 80px;
+		width: 100%;
+		height: 90vh;
+		background-image: url("../../images/face4.jpg");
+		background-position: center;
+		background-repeat: no-repeat;
+		background-size: cover;
+	}
+	.filter{
+		width: 100%;
+		height: 100%;
+		background-color: rgba(0,0, 0, 0.1);
+		display: flex;
+		flex-direction: column;
+		align-items: center;
+	}
+	.filter h2{
+		font-weight: 100;
+		font-size: 20px;
+		text-shadow: 1px 1px 1px black;
+		text-transform: uppercase;
+	}
+	.filter h2 span{
+		font-size: 30px;
+	}
+	.filter .h2-1{
+		margin-top: 20px;
+		overflow: hidden;
+		background: linear-gradient(90deg, #00d8ff, silver, #00d8ff);
+		background-repeat: no-repeat;
+		background-size: 40%;
+		background-clip: text;
+		-webkit-background-clip: text;
+		-webkit-text-fill-color: rgba(250, 235, 215, 0.5);
+		animation: shine 2s cubic-bezier(1, 1, 0.22, 0.19) infinite,
+					reveal 2s ease-in;
+		-moz-animation: mozGlow 2s linear infinite alternate 2s;
+		letter-spacing: 2px;
+	}
+	.filter .h2-2{
+		overflow: hidden;
+		background: linear-gradient(90deg, #00d8ff,silver, #00d8ff);
+		background-repeat: no-repeat;
+		background-size: 40%;
+		background-clip: text;
+		-webkit-background-clip: text;
+		-webkit-text-fill-color: rgba(250, 235, 215, 0.5);
+		animation: reverseShine 2s cubic-bezier(1, 1, 0.22, 0.19) infinite,
+					reveal 2s ease-in;
+		-moz-animation: mozGlow 2s linear infinite alternate 2s;
+		letter-spacing: 2px;
+		margin: 0;
+	}
+
+	.button-about-me{
+		margin-top: 325px;
+		padding: 20px;
+		font-size: 18px;
+		border: 2px solid antiquewhite;	
+		background-color: rgba(0, 0, 0, 0.4);
+		border-radius: 5px;
+		transition: all 0.1s linear;
+		color: antiquewhite;
+		text-decoration: none;
+		white-space: nowrap;
+	}
+}
+
+@media only screen and (min-width: 425px) {
+	.section-1{
+		height: 600px;
+	}
+}
+
+@media only screen and (min-width: 1024px) {
+	.section-1 {
+		height: 1000px;
+		background-image: url("../../images/face9.jpg");
+	}
+	.filter h2:first-of-type{
+		margin-top: 50px !important;
+	}
+	.section-1 h2{
+		font-size: 25px;
+		font-weight: 300;	
+	}
+	.button-about-me{
+		margin-top: 250px;
+	}
+	.button-about-me:hover{
+		background-color: antiquewhite;
+		border-color: antiquewhite;
+		border-radius: 20px/25%;
+		color: black;
+		font-weight: 500;
+		letter-spacing: 1.4px;
+	}
+}
+
+@media only screen and (min-width: 1440px){
+	.filter h2 {
+		font-size: 40px;
+	}
+	.filter h2 span{
+		font-size: 50px;
+	}
+}
+
+@media only screen and (min-width: 1800px){
+	.section-1 {
+		height: 1300px;
+	}
+	.section-1 .h2-1 span{
+		font-size: 70px;
+	}
+	.section-1 .h2-2{
+		font-size: 50px;
+	}
+	.button-about-me {
+		font-size: 30px;
+		padding: 30px;
+		margin-top: 350px;
+	}
+	.button-about-me:hover{
+		border-radius: 25px/25%;
+	}
+}
+
+@media only screen and (min-width: 2560px){
+	.section-1{
+		height: 1500px;
+	}
+}
+
+
+@media only screen and (min-width: 4000px){
+	.section-1 {
+		height: 3000px;
+	}
+	.button-about-me{
+		margin-top: 1300px;
+	}
+	.button-aboutMe:hover{
+		border-radius: 10%/40px;
+	}
+}
