+++ src/components/section-skills/skill-card/skill-card.component.css
